arch: some fixes

+++ scripts/bin/services/keyd/init.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env sh
 
-sudo rm /etc/keyd/default.conf
+[ -f /etc/keyd/default.conf ] && sudo rm /etc/keyd/default.conf
 
 sudo ln -s ~/bin/services/keyd/default.conf /etc/keyd/
 
-sudo service keyd restart
+sudo systemctl enable keyd
+~/bin/services/keyd/restart-keyd
